[ConfigBundle] Fixed and replaced not working validation in ConfigCompilerPass
