+++ src/Kunstmaan/ConfigBundle/DependencyInjection/Configuration.php
@@ -4,6 +4,7 @@ namespace Kunstmaan\ConfigBundle\DependencyInjection;
 
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
+use Kunstmaan\ConfigBundle\Entity\ConfigurationInterface as KunstmaanConfigurationInterface;
 
 /**
  * This is the class that validates and merges configuration from your app/config files
@@ -28,10 +29,27 @@ class Configuration implements ConfigurationInterface
         $rootNode
             ->children()
                 ->arrayNode('entities')
-                    ->defaultValue(array())
+                    ->defaultValue([])
                     ->normalizeKeys(false)
                     ->info('The list of entities to manage in the settings zone')
                     ->prototype('scalar')->end()
+                    ->beforeNormalization()
+                        ->always()
+                        ->then(function ($entities) {
+                            foreach ($entities as $entity) {
+                                if (!class_exists($entity)) {
+                                    throw new \InvalidArgumentException(sprintf('Entity "%s" does not exist', $entity));
+                                }
+
+                                // Check if entity implements the ConfigurationInterface.
+                                if (!in_array(KunstmaanConfigurationInterface::class, class_implements($entity))) {
+                                    throw new \RuntimeException(sprintf('The entity class "%s" needs to implement the %s', $entity, KunstmaanConfigurationInterface::class));
+                                }
+                            }
+
+                            return $entities;
+                        })
+                    ->end()
                 ->end()
             ->end();
