+++ src/Kunstmaan/ConfigBundle/KunstmaanConfigBundle.php
@@ -3,8 +3,6 @@
 namespace Kunstmaan\ConfigBundle;
 
 use Kunstmaan\ConfigBundle\DependencyInjection\Compiler\DeprecateClassParametersPass;
-use Kunstmaan\ConfigBundle\DependencyInjection\Compiler\KunstmaanConfigConfigurationPass;
-use Symfony\Component\DependencyInjection\Compiler\PassConfig;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
@@ -12,7 +10,6 @@ class KunstmaanConfigBundle extends Bundle
 {
     public function build(ContainerBuilder $container)
     {
-        $container->addCompilerPass(new KunstmaanConfigConfigurationPass(), PassConfig::TYPE_BEFORE_REMOVING);
         $container->addCompilerPass(new DeprecateClassParametersPass());
     }
 }
